Model side almost finished and Course Edit screen
The model side was done, with their Sqlite implementations, a few tests
were done, and it seems working. A Course Edit screen was done to add more
courses, a dependecy on attendances was created: every time a course is
manipulated (deleted, updated, or created), AttendanceManager is notified
and further notify its provider observers. In the case of creation, a new
Attendance is even created.