Card controls
Added events to card controls, now they increase and decrease the
absences bar. Added a isolated component to handle trials and retrials in
order not to trigger a couple of db operations in a row. Instead, if the
trials (the user pressing the buttons) are too fast, they trigger one single
trial after some defined interval.