Bugfixes on card controls and adjustments
It was recreating the discrete progress bar every time because of
wrap_content in layout_height of the list view, which resulted in
multiple calls to getView() of the adapter. That was fixed. The
listener for changes on the Attendance object of the card now is
set to update the database.