Preferences skeleton created and some clean up
Preferences stubs created for all the facets (read later) and one
preference added for Attendance (the total limit multiplier). Along with
preferences, it was created a custom Preference, namely
EditDecimalPreference, for editing preferences represented by decimal
values in a dialog, and also the Summarizer, which walks the preferences
tree and replaces a previously defined placeholder in the summary by the
current value for that preference. The cleaning up part refers to some
utility classes having been prefixed the keyword final, and a new class,
Facets, meant to be for static reference throughout the code and to
centralize aspects and resources of the current 3 facets of the
application: Attendance, Calendar and Report.