Fixes absence badge alignment
Before this diff, if the absence count for some course went from X to X.5, the
absence bar would readjust itself because of the new width. With this diff the
width is estimated as the bigger one, centralizing the text, thus the user won't
be disturbed by a resizing absence bar.