Adjust keyboard positioning when editing/creating course + nits
Put the keyboard above the cards view when editing a course or creating a new
one. Besides, display the keyboard instantaneously when the user goes to the edit
view instead of waiting for the user to fucus on the EditText. Some magic
numbers on the same class (CourseEditFragment) were converted to constants.