Animations when editing/creating card and suppress db update system
Created a mechanism for canceling view updates triggered by the db listener that
is used when the user is increasing absences on a card (before it also worked,
because there was a comparison of the models that came from the server and the
current ones, and since the update also modified the current Absence model, it
interpreted there was no changes), and it's supposed to be used when the view
triggers a update of the db but also updates itself needing no further changes
on the ui. Also created animations using the rebound library for spring
simulations when the user creates a new card or updates an existing one. This is
handled by the adapter, thus, any update/creation will be animated. This is
implemented tracking the changes between the old list of models and the new one,
and animating each action (creation, update, deletion) accordingly.